requirements.txt
Pillow==9.3.0
PySide2==5.15.2.1